full flow of conversation with bot work, prompts need to be revised
