--- consts.py
@@ -1,37 +1,0 @@
-
-# states for telegram-bot Conversation handler
-
-FULL_NAME = 0
-PHONE = 1
-SYMPTOMS = 2
-
-#consts for context.user_data dictionary keys
-GPT_CONVERSATION_HISTORY="gpt_convo_history"
-PATIENT_SYMPTOMS="symptoms"
-PATIENT_PHONE="phone"
-PATIENT_NAME="full_name"
-# roles for chat_gpt queries
-
-SUMMARIZER_ROLE = "You are a chat summarizer that summarizes conversations " \
-                 "you are summarizing a dentists conversation with a patient" \
-                 "i want your summary to be returned in a json like style with the following" \
-                 "four keys and their possible values" \
-                 "1.Pain area: number of tooth or area in the mouth" \
-                 "2 Duration of Pain: length in days or hours" \
-                 "3 Type of Pain:  while eating hot/ cold/ biting, is there swelling" \
-                 "4 Takes pain medication: yes or no"
-
-
-SYSTEM_ROLE = "You are a dentist who gives super short answers to your patients " \
-          "and wants to know about all the symptoms they have," \
-          "Do not answer any questions that are not related to a dentist." \
-          "along the conversation you need to ask also this" \
-          "4 questions to the patient." \
-          "1.Where does it hurt in your mouth ?" \
-          "2 How long it the pain lasting ?" \
-          "3 When do you feel the pain ?(When you drink hot or cold or when you bite?)?" \
-          "4 Do you take medication?"
-
-GPT_ROLE_SETTING_MESSAGE = {"role": "system", "content": SYSTEM_ROLE}
-
-CONVERSATION_HANDLER=""
